+++ src/main/java/ovh/angrysoft/homedaemon/devices/traits/occupancy/OccupancySensing.java
@@ -1,4 +1,4 @@
-package ovh.angrysoft.homedaemon.devices.traits;
+package ovh.angrysoft.homedaemon.devices.traits.occupancy;
 
 public interface OccupancySensing {
     public boolean getOccupancy();