+++ src/main/java/ovh/angrysoft/homedaemon/devices/xiaomi/aqara/SensorMotionAq2.java
@@ -5,7 +5,7 @@ import ovh.angrysoft.homedaemon.devices.DeviceAttribute;
 import ovh.angrysoft.homedaemon.devices.DeviceInfo;
 import ovh.angrysoft.homedaemon.devices.ZigbeeBaseDevice;
 import ovh.angrysoft.homedaemon.devices.traits.Illuminance;
-import ovh.angrysoft.homedaemon.devices.traits.OccupancySensing;
+import ovh.angrysoft.homedaemon.devices.traits.occupancy.OccupancySensing;
 import ovh.angrysoft.homedaemon.exceptions.attributes.AttributeAlreadyExist;
 
 public class SensorMotionAq2 extends ZigbeeBaseDevice implements OccupancySensing, Illuminance {